# docs/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "短剧API",
    "description": "短剧平台API接口文档，支持用户认证、视频管理、观看进度等功能",
    "version": "1.0.1",
    "contact": {
      "name": "API支持",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.example.com",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "VideoItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "视频ID"
          },
          "title": {
            "type": "string",
            "description": "视频标题"
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图片URL"
          },
          "playCount": {
            "type": "integer",
            "description": "播放次数"
          },
          "score": {
            "type": "number",
            "description": "评分"
          },
          "isHot": {
            "type": "boolean",
            "description": "是否热门"
          },
          "isVip": {
            "type": "boolean",
            "description": "是否VIP"
          }
        }
      },
      "VideoDetailInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "视频ID"
          },
          "title": {
            "type": "string",
            "description": "视频标题"
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图片URL"
          },
          "description": {
            "type": "string",
            "description": "视频描述"
          },
          "starring": {
            "type": "string",
            "description": "主演名单，多个演员用逗号分隔"
          },
          "actor": {
            "type": "string",
            "description": "完整演员名单，多个演员用逗号分隔"
          },
          "director": {
            "type": "string",
            "description": "导演信息，多个导演用逗号分隔"
          },
          "score": {
            "type": "string",
            "description": "评分"
          },
          "playCount": {
            "type": "integer",
            "description": "播放次数"
          },
          "serialCount": {
            "type": "integer",
            "description": "总集数"
          },
          "updateStatus": {
            "type": "string",
            "description": "更新状态"
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EpisodeInfo"
            }
          }
        }
      },
      "EpisodeInfo": {
        "type": "object",
        "properties": {
          "episodeId": {
            "type": "integer",
            "description": "集数ID"
          },
          "title": {
            "type": "string",
            "description": "集数标题"
          },
          "episodeTitle": {
            "type": "string",
            "description": "集数名称"
          },
          "duration": {
            "type": "integer",
            "description": "时长（秒）"
          },
          "isVip": {
            "type": "boolean",
            "description": "是否VIP"
          },
          "isLast": {
            "type": "boolean",
            "description": "是否最后一集"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/video/progress": {
      "post": {
        "tags": ["视频"],
        "summary": "保存观看进度",
        "description": "保存用户的视频观看进度",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "episodeId": {
                    "type": "integer",
                    "description": "剧集ID"
                  },
                  "stopAtSecond": {
                    "type": "integer",
                    "description": "停止观看的秒数"
                  }
                },
                "required": ["episodeId", "stopAtSecond"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存成功"
          }
        }
      },
      "get": {
        "tags": ["视频"],
        "summary": "获取观看进度",
        "description": "获取用户的视频观看进度",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "episodeId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "剧集ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stopAtSecond": {
                      "type": "integer",
                      "description": "停止观看的秒数"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/video/comment": {
      "post": {
        "tags": ["视频"],
        "summary": "发表评论或弹幕",
        "description": "为视频剧集发表评论或弹幕",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "episodeId": {
                    "type": "integer",
                    "description": "剧集ID"
                  },
                  "content": {
                    "type": "string",
                    "description": "评论内容"
                  },
                  "appearSecond": {
                    "type": "integer",
                    "description": "弹幕出现时间（秒），普通评论可不填"
                  }
                },
                "required": ["episodeId", "content"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发表成功"
          }
        }
      }
    },
    "/api/video/details": {
      "get": {
        "tags": ["视频"],
        "summary": "获取视频详情",
        "description": "获取视频的详细信息，包括剧集列表、演员信息、导演信息等",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "视频ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "detailInfo": {
                          "$ref": "#/components/schemas/VideoDetailInfo"
                        },
                        "like": {
                          "type": "object",
                          "properties": {
                            "count": {
                              "type": "integer"
                            },
                            "selected": {
                              "type": "boolean"
                            }
                          }
                        },
                        "favorites": {
                          "type": "object",
                          "properties": {
                            "count": {
                              "type": "integer"
                            },
                            "selected": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    },
                    "msg": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/video/media": {
      "get": {
        "tags": ["视频"],
        "summary": "获取用户媒体列表",
        "description": "获取用户的媒体内容列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "分类ID"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["short", "series"]
            },
            "description": "媒体类型"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoItem"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/public/video/media": {
      "get": {
        "tags": ["公共视频"],
        "summary": "获取公共媒体列表",
        "description": "获取公共媒体内容列表",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "分类ID"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["short", "series"]
            },
            "description": "媒体类型"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["latest", "like", "play"],
              "default": "latest"
            },
            "description": "排序方式"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "页码"
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoItem"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}